Additional Unit Tests, OssBitString deprecation, and OssPathHistory refactor

+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/J2735PivotPointDescription.java
@@ -7,7 +7,7 @@ import us.dot.its.jpo.ode.plugin.asn1.Asn1Object;
 public class J2735PivotPointDescription extends Asn1Object {
 	private static final long serialVersionUID = 1L;
 
-	public Integer pivotOffset;
+	public BigDecimal pivotOffset;
 	public BigDecimal pivotAngle;
 	public Boolean pivots;
